refactor data handling; update ID types to Long and implement RandomAccessManager for efficient record management

+++ UniDB/src/main/java/Storage/Collection.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface Collection {
     boolean insertOne(Student student);
-    boolean deleteOne(int id);
-    Student findByID(int id);
+    boolean deleteOne(Long id);
+    Student findByID(Long id);
     List<Student> findAll();
 }
